feat: 모집안내 탭 헤더 컴포넌트 구현
